requirements.txt
google-genai
PyAudio
fastapi
uvicorn
loguru
python-dotenv